requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests>=2.31.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0